Consolidate Binospec unit tests into test_mmt_binospec.py
Merge the five separate Binospec test files (test_binospec_bpm,
test_binospec_ifu_extract, test_binospec_overscan, test_binospec_setup, and
the new test_mmt_binospec_comb) into a single test_mmt_binospec.py, grouped
by topic and named to match the spectrograph module.  Add explanatory
messages to every combination-group assert.  Per review feedback.
